offender checkin integration tests
Also includes:
- change cascade type on some entities to allow clearing the DB
  in tests
- add validation to few routes
- fix: properly set manual id check result

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/esupervisionapi/utils/S3UploadService.kt
@@ -52,7 +52,7 @@ data class CheckinVideoKey(
 
 data class CheckinPhotoKey(
   val checkin: UUID,
-  val index: Long,
+  val index: Int,
 ) : S3Keyable()
 
 @Service
@@ -79,7 +79,7 @@ class S3UploadService(
   /**
    * We can have multiple checkin photo objects, hence the index param
    */
-  private fun putObjectRequest(checkin: OffenderCheckin, contentType: String, index: Long): PutObjectRequest = putObjectRequest(videoUploadBucket, CheckinPhotoKey(checkin.uuid, index).toKey(), contentType)
+  private fun putObjectRequest(checkin: OffenderCheckin, contentType: String, index: Int): PutObjectRequest = putObjectRequest(videoUploadBucket, CheckinPhotoKey(checkin.uuid, index).toKey(), contentType)
 
   /**
    * Generates a pre-signed URL for uploading a file to S3.
@@ -126,7 +126,7 @@ class S3UploadService(
   fun generatePresignedUploadUrl(
     checkin: OffenderCheckin,
     contentType: String = "application/octet-stream",
-    index: Long,
+    index: Int,
     duration: Duration,
   ): URL {
     val putRequest = putObjectRequest(checkin, contentType, index)
